move dev tasks to own file
